Text-system and more!
I'm finally done with my text-rendering system, and at the same time I
got to improve the rendering-system that was already in place.

I've also started setting up a hierarchy-based entity-system.
And on top of all this I've also fixed the aspect-adjustment on the
windowing-system, and made the error-handler make a system-dependent
beep-sound on all notice-errors.