More menuing!
Added the option to reset the game in the pause-menu. This also happens
automatically when the player receives a "game over".
The player can now view the keyboard-controls and XBOX-controls for the
game in the pause-menu.
Added movement to the enemies.
Also playing with a gamepad now actually works :P