Levels and patterns!
I've set up a level-system that handles the music and the backgrounds
for different levels, but most importantly does it also manage a "waves
of enemies"-system, where the game spawns in the next enemies when the
player kills the current enemies.

I've also started experimenting with bullet patterns, and different
attacks!